banking-intent-unsloth/requirements.txt
unsloth
trl
peft
accelerate
bitsandbytes
datasets
pandas
pyyaml
torch
transformers
tqdm